BROKEN 4/153: Map local gensyms to stable server gensyms
The rdf-spec specifications require us to treat arbitrary blank node IDs
as stable for the lifetime of a connection.  For example, '_:genid1'
must always point to the same node.  Unfortunately, AllegroGraph only
guarantees the stability of blank node IDs that we explicitly allocate.

To fix this, we need to map between client-side IDs and server-side IDs
in our adapter, and allocate server-side IDs as needed.