Unlicense this software
An e-mail from my client contained: "For this project I don't think there
is a real preference for the licenses.  I would say lets use the unlicense
if that is what the community prefers."

Using UNLICENSE for this project will allow us to work more easily with
upstream RDF.rb.