Optionally infer triples using AllegroGraph's RDFS++ support
Aymeric Brisse requested support for AllegroGraph's RDFS++ inference,
and supplied me with example RDF data and several test cases.  Those
test cases made it easy to write this patch.

This patch adds support for "query options", which can be passed to
build_query, and which are interpreted by raw_query.  Right now, the
only supported query option is 'infer', which controls whether or not we
use AllegroGraph's inference support.

To add new query options, we need to update 3 places in the code:

  query_spec.rb: Add a new, self-contained spec.
  AbstractRepository#build_query: Document the option.
  AbstractRepository#raw_query: Implement the option.